requirements.txt
pandas
requests
python-dotenv
psycopg2-binary
alembic
asyncpg
SQLAlchemy  